fix arm 3d view sometimes showing last selected arm config and new
there were several issues
- animation frame loop was not being canceled between changes
- loadStl files needs to finish before rendering the scene
